backend/requirements.txt
aiohttp==3.6.2
alembic==1.4.3
async-timeout==3.0.1
attrs==20.2.0
bcrypt==3.2.0
boto3==1.14.62
boto3-stubs==1.14.62.0
botocore==1.17.62
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
docutils==0.16
ecdsa==0.16.0
email-validator==1.0.5
fastapi==0.61.1
future==0.18.2
gunicorn==20.0.4
h11==0.10.0
httptools==0.1.1
idna==2.9
jmespath==0.9.5
Mako==1.1.2
MarkupSafe==1.1.1
memory-profiler==0.57.0
multidict==4.7.5
passlib==1.7.2
psutil==5.7.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
PyMuPDF==1.17.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
python-multipart==0.0.5
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.19
starlette==0.13.6
tenacity==6.1.0
urllib3==1.25.8
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
yarl==1.4.2
